+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 Under development
 
 * Make a slight tweak to the output text so that it sounds gentler
+* Give the supported remainder options in the error output ([#12](https://github.com/funnelfiasco/coffee_match/issues/12))
 
 # version 0.2
 Released 2 April 2020